Separate manager implementation by modules and subs
